+++ src/utils/checksum.c
@@ -19,7 +19,7 @@ uint32_t _z_crc32(const uint8_t *message, size_t len) {
     for (size_t i = 0; i < len; i++) {
         crc = crc ^ (uint32_t)message[i];
         for (int j = 0; j < 8; j++) {
-            crc = (crc >> (uint32_t)1) ^ ((uint32_t)0x04C11DB7 & (-(crc & (uint32_t)1)));
+            crc = (crc >> 1) ^ ((uint32_t)0x04C11DB7 & (-(crc & (uint32_t)1)));
         }
     }
     return ~crc;
